Refactor pkg name
