+++ pkg/util/azureclient/applens/applens_client_test.go
@@ -16,7 +16,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 
-	testhttp "github.com/Azure/ARO-RP/test/util/http"
+	testhttp "github.com/Azure/ARO-RP/test/util/http/server"
 )
 
 func TestEnsureErrorIsGeneratedOnResponse(t *testing.T) {
