+++ test/util/http/server/server_mock_test.go
@@ -1,4 +1,4 @@
-package applens
+package server
 
 // Copyright (c) Microsoft Corporation.
 // Licensed under the Apache License 2.0.
